🎨 Improved structure of 'types'
