+++ src/types/feature/index.ts
@@ -1,0 +1,4 @@
+export type Feature = {
+  id: number;
+  name: string;
+};
