+++ src/types/recipeBook/category.ts
@@ -1,0 +1,6 @@
+export type Category = {
+  id: number;
+  name: string;
+  createdAt: string;
+  updatedAt: string;
+};
