+++ src/types/recipeBook/origin.ts
@@ -1,0 +1,6 @@
+export type Origin = {
+  id: number;
+  name: string;
+  createdAt: string;
+  updatedAt: string;
+};
