+++ src/types/recipeBook/step.ts
@@ -1,0 +1,4 @@
+export type Step = {
+  number: number;
+  instruction: string;
+};
