+++ src/types/search/index.ts
@@ -1,0 +1,3 @@
+export type SearchForm = {
+  searchText: string;
+};
